chore: update package dependencies and enhance functionality
- Upgraded various dependencies in package.json files, including "@types/aws-lambda", "@types/node", and AWS SDK packages to their latest versions for improved performance and security.
- Updated the 'docker-build' version in sst.config.ts for better compatibility.
- Refined the daily-devotion.ts file to handle web push errors more effectively by checking for specific error instances.
- Enhanced the bible reading system prompt in the AI package to prevent repetition and improve output formatting.
- Made minor updates to other packages, including 'hono', 'svix', and 'react-markdown', ensuring all components are up-to-date.

+++ apps/functions/src/jobs/daily-devotion.ts
@@ -9,7 +9,7 @@ import { wrapHandler } from '@sentry/aws-serverless';
 import { formatDate } from 'date-fns';
 import { eq } from 'drizzle-orm';
 import { Resource } from 'sst';
-import webPush from 'web-push';
+import webPush, { WebPushError } from 'web-push';
 
 webPush.setVapidDetails(
   'mailto:support@theaistudybible.com',
@@ -74,7 +74,7 @@ const sendPushNotifications = async (devotion: Devotion) => {
         )
         .catch(async (error) => {
           console.error(error);
-          if (error.statusCode === 410) {
+          if (error instanceof WebPushError && error.statusCode === 410) {
             await db.delete(pushSubscriptions).where(eq(pushSubscriptions.id, subscription.id));
           }
         }),
